First idea, originally from 07/Dec/2017
After doing this test, the website provides an extensive explanation of what people that identify with that personality share in common. It is pretty similar to a zodiac sign explanation, but with actual evidence.

The test uses the bayes classifier to identify for each variable measured (mind,energy,nature,tactics) what class you are more similar to, in other words it takes the characteristic that exceeds 50 percent and sets as THE ONE you identify with. The problem is that probably no one identifies 100% with any of the characteristics, so I decided to actually calculate the probability distribution of identification and it provided interesting interpretations for myself.